test(validator): split large test files
